+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v3.0.0 _14 July 2022_
 
-- Added a mock mode
+- Added a mock mode enabled by MOCK_MODE=true
 
 ## v2.0.0 _17 June 2022_